requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==3.4.8
defusedxml==0.7.1
Django==2.1.5
django-adminlte3==0.1.6
django-allauth==0.44.0
django-bootstrap-modal-forms==2.1.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-oauth2==3.2
filebrowser-safe==1.1.1
grappelli-safe==1.1.1
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.0.1
PyMySQL==1.0.2
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
shortuuid==1.0.11
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.13
